Refactor internal hub api
