+++ cmd/chart-tracker/errors.go
@@ -5,6 +5,7 @@ import (
 	"strings"
 	"sync"
 
+	"github.com/artifacthub/hub/internal/api"
 	"github.com/artifacthub/hub/internal/hub"
 	"github.com/rs/zerolog/log"
 )
@@ -20,14 +21,14 @@ const (
 // collected errors can be flushed, which will store them in the database.
 type errorsCollector struct {
 	ctx    context.Context
-	hubAPI *hub.Hub
+	hubAPI *api.API
 
 	mu     sync.Mutex
 	errors map[string][]error // K: chart repository id
 }
 
 // newErrorsCollector creates a new errorsCollector instance.
-func newErrorsCollector(ctx context.Context, hubAPI *hub.Hub, repos []*hub.ChartRepository) *errorsCollector {
+func newErrorsCollector(ctx context.Context, hubAPI *api.API, repos []*hub.ChartRepository) *errorsCollector {
 	ec := &errorsCollector{
 		ctx:    ctx,
 		hubAPI: hubAPI,
@@ -58,7 +59,7 @@ func (c *errorsCollector) flush() {
 			errStr.WriteString(err.Error())
 			errStr.WriteString("\n")
 		}
-		err := c.hubAPI.SetChartRepositoryLastTrackingResults(c.ctx, chartRepositoryID, errStr.String())
+		err := c.hubAPI.ChartRepositories.SetLastTrackingResults(c.ctx, chartRepositoryID, errStr.String())
 		if err != nil {
 			log.Error().Err(err).Str("repoID", chartRepositoryID).Send()
 		}
