+++ cmd/chart-tracker/worker.go
@@ -13,6 +13,7 @@ import (
 	"sync"
 	"time"
 
+	"github.com/artifacthub/hub/internal/api"
 	"github.com/artifacthub/hub/internal/hub"
 	"github.com/artifacthub/hub/internal/img"
 	"github.com/rs/zerolog"
@@ -26,14 +27,14 @@ type worker struct {
 	ctx        context.Context
 	id         int
 	ec         *errorsCollector
-	hubAPI     *hub.Hub
+	hubAPI     *api.API
 	imageStore img.Store
 	logger     zerolog.Logger
 	httpClient *http.Client
 }
 
 // newWorker creates a new worker instance.
-func newWorker(ctx context.Context, id int, ec *errorsCollector, hubAPI *hub.Hub, imageStore img.Store) *worker {
+func newWorker(ctx context.Context, id int, ec *errorsCollector, hubAPI *api.API, imageStore img.Store) *worker {
 	return &worker{
 		ctx:        ctx,
 		id:         id,
@@ -170,7 +171,7 @@ func (w *worker) handleJob(j *job) error {
 	}
 
 	// Register package
-	err = w.hubAPI.RegisterPackage(w.ctx, p)
+	err = w.hubAPI.Packages.Register(w.ctx, p)
 	if err != nil {
 		w.ec.append(
 			j.repo.ChartRepositoryID,
