+++ internal/user/tmpl_email_verification.go
@@ -1,8 +1,8 @@
-package hub
+package user
 
 import "html/template"
 
-var organizationInvitationTmpl = template.Must(template.New("").Parse(`
+var emailVerificationTmpl = template.Must(template.New("").Parse(`
 <!doctype html>
 <html>
   <head>
@@ -100,7 +100,7 @@ var organizationInvitationTmpl = template.Must(template.New("").Parse(`
           <div class="content" style="box-sizing: border-box; display: block; Margin: 0 auto; max-width: 580px; padding: 10px;">
 
             <!-- START CENTERED WHITE CONTAINER -->
-            <span class="preheader" style="color: transparent; display: none; height: 0; max-height: 0; max-width: 0; opacity: 0; overflow: hidden; mso-hide: all; visibility: hidden; width: 0;">Invitation to {{ .orgName }} organization on Artifact Hub</span>
+            <span class="preheader" style="color: transparent; display: none; height: 0; max-height: 0; max-width: 0; opacity: 0; overflow: hidden; mso-hide: all; visibility: hidden; width: 0;">Welcome to Artifact Hub!</span>
             <table class="main" style="border-collapse: separate; mso-table-lspace: 0pt; mso-table-rspace: 0pt; width: 100%; background: #ffffff; border-radius: 3px; border-top: 7px solid #659DBD;">
 
               <!-- START MAIN CONTENT AREA -->
@@ -110,7 +110,7 @@ var organizationInvitationTmpl = template.Must(template.New("").Parse(`
                     <tr>
                       <td style="font-family: sans-serif; font-size: 14px; vertical-align: top;">
                         <p style="font-family: sans-serif; font-size: 14px; font-weight: normal; margin: 0; Margin-bottom: 15px;">Hi!</p>
-                        <p style="font-family: sans-serif; font-size: 14px; font-weight: normal; margin: 0; Margin-bottom: 30px;">You have been invited to join <b>{{ .orgName }}</b> organization on Artifact Hub.</p>
+                        <p style="font-family: sans-serif; font-size: 14px; font-weight: normal; margin: 0; Margin-bottom: 30px;">Welcome to Artifact Hub! You are only one step from being able to sign in on our site. Please simply click on the link below to confirm your account.</p>
                         <table border="0" cellpadding="0" cellspacing="0" class="btn btn-primary" style="border-collapse: separate; mso-table-lspace: 0pt; mso-table-rspace: 0pt; width: 100%; box-sizing: border-box;">
                           <tbody>
                             <tr>
@@ -118,7 +118,7 @@ var organizationInvitationTmpl = template.Must(template.New("").Parse(`
                                 <table border="0" cellpadding="0" cellspacing="0" style="border-collapse: separate; mso-table-lspace: 0pt; mso-table-rspace: 0pt; width: auto;">
                                   <tbody>
                                     <tr>
-                                      <td style="font-family: sans-serif; font-size: 14px; border-radius: 5px; vertical-align: top; text-align: center;"> <a href="{{ .link }}" target="_blank" style="display: inline-block; color: #ffffff; background-color: #39596C; border: solid 1px #39596C; border-radius: 5px; box-sizing: border-box; cursor: pointer; text-decoration: none; font-size: 14px; font-weight: bold; margin: 0; padding: 12px 25px; text-transform: capitalize; border-color: #39596C;">Accept invitation</a> </td>
+                                      <td style="font-family: sans-serif; font-size: 14px; border-radius: 5px; vertical-align: top; text-align: center;"> <a href="{{ .link }}" target="_blank" style="display: inline-block; color: #ffffff; background-color: #39596C; border: solid 1px #39596C; border-radius: 5px; box-sizing: border-box; cursor: pointer; text-decoration: none; font-size: 14px; font-weight: bold; margin: 0; padding: 12px 25px; text-transform: capitalize; border-color: #39596C;">Confirm your account</a> </td>
                                     </tr>
                                   </tbody>
                                 </table>
@@ -130,12 +130,13 @@ var organizationInvitationTmpl = template.Must(template.New("").Parse(`
                           <tbody>
                             <tr>
                               <td class="content-block powered-by" style="font-family: sans-serif; vertical-align: top; font-size: 11px; color: #545454; padding-bottom: 30px; padding-top: 10px;">
-                                <p style="color: #545454; font-size: 11px; text-decoration: none;">You can also accept the invitation by visiting the page directly at <span style="color: #545454; background-color: #ffffff;">{{ .link }}</span></p>
+                                <p style="color: #545454; font-size: 11px; text-decoration: none;">Or you can copy-paste this link: <span style="color: #545454; background-color: #ffffff;">{{ .link }}</span></p>
                               </td>
                             </tr>
                           </tbody>
                         </table>
-                        <p style="font-family: sans-serif; font-size: 14px; font-weight: normal; margin: 0; Margin-bottom: 15px;">Thanks.</p>
+                        <p style="font-family: sans-serif; font-size: 14px; font-weight: normal; margin: 0; Margin-bottom: 15px;">After activation you may sign in to Artifact Hub using your credentials.</p>
+                        <p style="font-family: sans-serif; font-size: 14px; font-weight: normal; margin: 0; Margin-bottom: 15px;">Thanks for creating an account.</p>
                       </td>
                     </tr>
                   </table>
@@ -150,7 +151,7 @@ var organizationInvitationTmpl = template.Must(template.New("").Parse(`
               <table border="0" cellpadding="0" cellspacing="0" style="border-collapse: separate; mso-table-lspace: 0pt; mso-table-rspace: 0pt; width: 100%;">
                 <tr>
                   <td class="content-block powered-by" style="font-family: sans-serif; vertical-align: top; padding-bottom: 10px; padding-top: 10px; font-size: 10px; color: #545454; text-align: center;">
-                    <p style="color: #545454; font-size: 10px; text-align: center; text-decoration: none;">If this email means nothing to you, then it is possible that somebody else has entered your user alias accidentally, so please ignore this email.</p>
+                    <p style="color: #545454; font-size: 10px; text-align: center; text-decoration: none;">Didn't create an Artifact Hub account? I's likely someone just typed in your email address by accident.<br>Feel free to ignore this email.</p>
                   </td>
                 </tr>
                 <tr>
